40k: Factor out common HTML code
Factor out common HTML code used by 40k renderers, so we can use it for KT. This includes table creation, options, and hideable behavior
